+++ tests/unit/asana/test_client.py
@@ -28,33 +28,24 @@ from tests.unit.asana import tester_data
 
 
 
-@pytest.fixture(name='raise_no_authorization_error')
-def fixture_raise_no_authorization_error():
+@pytest.fixture(name='raise_asana_error')
+def fixture_raise_asana_error(request):
     """
-    Returns a function that can be mocked over a call that simply forces it to
-    raise the asana.error.NoAuthorizationError.
+    Returns a function that can be used to mock a call, simply forcing it to
+    raise a marked `AsanaError` sub-error.  If no marker, will use a default
+    exception type.
     """
+    marker = request.node.get_closest_marker('asana_error_data')
+    if marker is None:
+        exception_type = asana.error.InvalidRequestError # Arbitrary
+    else:
+        exception_type = marker.args[0]
+
     def mock_raise(*args, **kwargs):
         """
         Simply raise the desired error.
         """
-        raise asana.error.NoAuthorizationError()
-
-    return mock_raise
-
-
-
-@pytest.fixture(name='raise_not_found_error')
-def fixture_raise_not_found_error():
-    """
-    Returns a function that can be mocked over a call that simply forces it to
-    raise the asana.error.NotFoundError.
-    """
-    def mock_raise(*args, **kwargs):
-        """
-        Simply raise the desired error.
-        """
-        raise asana.error.NotFoundError()
+        raise exception_type
 
     return mock_raise
 
@@ -92,35 +83,75 @@ def fixture_project_test():
 
 
 
-@pytest.fixture(name='section_in_project_test', scope='session')
-def fixture_section_in_project_test(project_test):
+@pytest.fixture(name='sections_in_project_test', scope='session')
+def fixture_sections_in_project_test(project_test):
     """
-    Creates a test section and returns the dict of data that should match the
-    'data' element returned by the API.
+    Creates some test sections and returns a list of them, each of which is the
+    dict of data that should match the 'data' element returned by the API.
 
-    Will delete the section once done with all tests.
+    Will delete the sections once done with all tests.
 
     This is not being used with the autouse keyword so that, if running tests
     that do not require this section fixture, they can run more optimally
     without the need to needlessly create and delete this section.  (Also,
     could not figure out how to get rid of all syntax and pylint errors).
 
-    ** Consumes 3 API calls. **
+    ** Consumes 5 API calls. **
+    (API call count is 2*num_sects + 1)
     """
     # pylint: disable=no-member     # asana.Client dynamically adds attrs
-    sect_name = tester_data._SECTION_TEMPLATE.substitute({'sid': uuid.uuid4()})
+    num_sects = 2
     client = aclient._get_client()
     me_data = aclient._get_me()
-    params = {
-        'name': sect_name,
-        'owner': me_data['gid'],
-    }
-    sect_data = client.sections.create_section_for_project(project_test['gid'],
-            params)
 
-    yield sect_data
+    sect_data_list = []
+    for _ in range(num_sects):
+        sect_name = tester_data._SECTION_TEMPLATE.substitute(
+                {'sid': uuid.uuid4()})
+        params = {
+            'name': sect_name,
+            'owner': me_data['gid'],
+        }
+        sect_data = client.sections.create_section_for_project(
+                project_test['gid'], params)
+        sect_data_list.append(sect_data)
 
-    client.sections.delete_section(sect_data['gid'])
+    yield sect_data_list
+
+    for i in range(num_sects):
+        client.sections.delete_section(sect_data_list[i]['gid'])
+
+
+
+def subtest_asana_error_handler_func(caplog, exception_type, log_index, func,
+        *args, **kwargs):
+    """
+    Executes a subtest to confirm `@asana_error_handler` is properly decorating
+    the given function.  This allows test for the given functions to setup
+    anything function-specific required prior to running these same tests steps.
+
+    Expected to be called by every function that uses the `@asana_error_handler`
+    decorator.
+
+    Args:
+      caplog (Caplog): The caplog fixture from the pytest test.
+      exception_type (AsanaError): The exact exception type that is expected to
+        be caught.  Can be as generic as `AsanaError`, but testing for something
+        more specific is better to improve coverage.
+      log_index (int): The index to check in caplog for the desired exception
+        log message.
+      func (function): The reference to the function to call to test.
+      *args ([any]): The positional arguments to pass to the function to test
+        `func`.
+      **kwargs ({str:any}): The keyword arguments to pass to teh function to
+        test `func`.
+    """
+    caplog.clear()
+    with pytest.raises(exception_type):
+        func(*args, **kwargs)
+    assert caplog.record_tuples[log_index][0] == 'asana_extensions.asana.client'
+    assert caplog.record_tuples[log_index][1] == logging.ERROR
+    assert 'API query failed' in caplog.messages[log_index]
 
 
 
@@ -139,6 +170,54 @@ def test_logging_capture_warnings(caplog):
     assert caplog.record_tuples[0][0] == 'py.warnings'
     assert caplog.record_tuples[0][1] == logging.WARNING
     assert 'Test warning' in caplog.record_tuples[0][2]
+
+
+
+def test_asana_error_handler(caplog):
+    """
+    Tests the `@asana_error_handler` decorator.
+    """
+    caplog.set_level(logging.ERROR)
+
+    def gen_text(text1, text2, text3):
+        return f'{text1} | {text2} | {text3}'
+
+    dec_gen_text = aclient.asana_error_handler(gen_text)
+    assert dec_gen_text('one', text3='three', text2='two') \
+            == 'one | two | three'
+    assert dec_gen_text._is_wrapped_by_asana_error_handler is True
+
+    def raise_error(exception_type):
+        raise exception_type
+
+    dec_raise_error = aclient.asana_error_handler(raise_error)
+    exception_types = [
+        asana.error.PremiumOnlyError,
+        asana.error.RateLimitEnforcedError,
+    ]
+    for exception_type in exception_types:
+        subtest_asana_error_handler_func(caplog, exception_type, 0,
+                dec_raise_error, exception_type)
+
+    assert dec_raise_error._is_wrapped_by_asana_error_handler is True
+
+
+
+@pytest.mark.parametrize('func_name', [
+    '_get_me',
+    'get_workspace_gid_from_name',
+    'get_project_gid_from_name',
+    'get_section_gid_from_name',
+    'get_user_task_list_gid',
+    'get_section_gids_in_project_or_utl',
+])
+def test_dec_usage_asana_error_handler(func_name):
+    """
+    Tests that functions that are expected to use the `@asana_error_handler`
+    decorator do in fact have it.
+    """
+    func = getattr(aclient, func_name)
+    assert func._is_wrapped_by_asana_error_handler is True
 
 
 
@@ -197,16 +276,11 @@ def test__get_me(monkeypatch, caplog):
             },
         }
 
-    caplog.clear()
+    # Function-specific practical test of @asana_error_handler
     monkeypatch.delattr(aclient._get_client, 'client')
     monkeypatch.setattr(config, 'read_conf_file', mock_read_conf_file)
-    with pytest.raises(asana.error.NoAuthorizationError):
-        aclient._get_me()
-    assert caplog.record_tuples == [
-            ('asana_extensions.asana.client', logging.ERROR,
-                "Failed to access API in _get_me() - Not Authorized:"
-                + " No Authorization: Not Authorized"),
-    ]
+    subtest_asana_error_handler_func(caplog, asana.error.NoAuthorizationError,
+            0, aclient._get_me)
 
 
 
@@ -265,8 +339,8 @@ def test__find_gid_from_name(caplog):
 
 
 
-def test_get_workspace_gid_from_name(monkeypatch, caplog,
-        raise_no_authorization_error):
+@pytest.mark.asana_error_data.with_args(asana.error.ForbiddenError)
+def test_get_workspace_gid_from_name(monkeypatch, caplog, raise_asana_error):
     """
     Tests the `get_workspace_gid_from_name()` method.
 
@@ -300,23 +374,17 @@ def test_get_workspace_gid_from_name(monkeypatch, caplog,
     assert 'name' in workspace
     assert 'resource_type' in workspace
 
+    # Function-specific practical test of @asana_error_handler
     # Need to monkeypatch cached client since class dynamically creates attrs
-    monkeypatch.setattr(client.workspaces, 'get_workspaces',
-            raise_no_authorization_error)
-
-    caplog.clear()
-    with pytest.raises(asana.error.NoAuthorizationError):
-        aclient.get_workspace_gid_from_name('one and only')
-    assert caplog.record_tuples == [
-            ('asana_extensions.asana.client', logging.ERROR,
-                "Failed to access API in get_workspace_gid_from_name() - Not"
-                + " Authorized: No Authorization"),
-    ]
+    monkeypatch.setattr(client.workspaces, 'get_workspaces', raise_asana_error)
+    subtest_asana_error_handler_func(caplog, asana.error.ForbiddenError, 0,
+            aclient.get_workspace_gid_from_name, 'one and only')
 
 
 
+@pytest.mark.asana_error_data.with_args(asana.error.NotFoundError)
 def test_get_project_gid_from_name(monkeypatch, caplog, project_test,
-        raise_no_authorization_error):
+        raise_asana_error):
     """
     Tests the `get_project_gid_from_name()` method.
 
@@ -355,23 +423,17 @@ def test_get_project_gid_from_name(monkeypatch, caplog, project_test,
     assert 'name' in project
     assert 'resource_type' in project
 
+    # Function-specific practical test of @asana_error_handler
     # Need to monkeypatch cached client since class dynamically creates attrs
-    monkeypatch.setattr(client.projects, 'get_projects',
-            raise_no_authorization_error)
-
-    caplog.clear()
-    with pytest.raises(asana.error.NoAuthorizationError):
-        aclient.get_project_gid_from_name(ws_gid, project_test['name'])
-    assert caplog.record_tuples == [
-            ('asana_extensions.asana.client', logging.ERROR,
-                "Failed to access API in get_project_gid_from_name() - Not"
-                + " Authorized: No Authorization"),
-    ]
+    monkeypatch.setattr(client.projects, 'get_projects', raise_asana_error)
+    subtest_asana_error_handler_func(caplog, asana.error.NotFoundError, 0,
+            aclient.get_project_gid_from_name, ws_gid, project_test['name'])
 
 
 
+@pytest.mark.asana_error_data.with_args(asana.error.InvalidTokenError)
 def test_get_section_gid_from_name(monkeypatch, caplog, project_test,
-        section_in_project_test, raise_no_authorization_error):
+        sections_in_project_test, raise_asana_error):
     """
     Tests the `get_section_gid_from_name()` method.
 
@@ -388,6 +450,9 @@ def test_get_section_gid_from_name(monkeypatch, caplog, project_test,
     """
     # pylint: disable=no-member     # asana.Client dynamically adds attrs
     caplog.set_level(logging.ERROR)
+
+    # Only need 1 section
+    section_in_project_test = sections_in_project_test[0]
 
     # Sanity check that this works with an actual section
     try:
@@ -409,24 +474,18 @@ def test_get_section_gid_from_name(monkeypatch, caplog, project_test,
     assert 'name' in section
     assert 'resource_type' in section
 
+    # Function-specific practical test of @asana_error_handler
     # Need to monkeypatch cached client since class dynamically creates attrs
     monkeypatch.setattr(client.sections, 'get_sections_for_project',
-            raise_no_authorization_error)
-
-    caplog.clear()
-    with pytest.raises(asana.error.NoAuthorizationError):
-        aclient.get_section_gid_from_name(project_test['gid'],
-                section_in_project_test['name'])
-    assert caplog.record_tuples == [
-            ('asana_extensions.asana.client', logging.ERROR,
-                "Failed to access API in get_section_gid_from_name() - Not"
-                + " Authorized: No Authorization"),
-    ]
+            raise_asana_error)
+    subtest_asana_error_handler_func(caplog, asana.error.InvalidTokenError, 0,
+            aclient.get_section_gid_from_name, project_test['gid'],
+            section_in_project_test['name'])
 
 
 
-def test_get_user_task_list_gid(monkeypatch, caplog,
-        raise_no_authorization_error, raise_not_found_error):
+@pytest.mark.asana_error_data.with_args(asana.error.ServerError)
+def test_get_user_task_list_gid(monkeypatch, caplog, raise_asana_error):
     """
     Tests the `get_user_task_list_gid()` method.
 
@@ -466,34 +525,19 @@ def test_get_user_task_list_gid(monkeypatch, caplog,
         aclient.get_user_task_list_gid(0, True, 0)
     assert 'Must provide `is_me` or `user_gid`, but not both.' in str(ex.value)
 
+    # Function-specific practical test of @asana_error_handler
     client = aclient._get_client()
     # Need to monkeypatch cached client since class dynamically creates attrs
     monkeypatch.setattr(client.user_task_lists, 'get_user_task_list_for_user',
-            raise_no_authorization_error)
-    caplog.clear()
-    with pytest.raises(asana.error.NoAuthorizationError):
-        aclient.get_user_task_list_gid(0, True)
-    assert caplog.record_tuples == [
-            ('asana_extensions.asana.client', logging.ERROR,
-                "Failed to access API in get_user_task_list_gid() - Not"
-                + " Authorized: No Authorization"),
-    ]
-
-    monkeypatch.setattr(client.user_task_lists, 'get_user_task_list_for_user',
-            raise_not_found_error)
-    caplog.clear()
-    with pytest.raises(asana.error.NotFoundError):
-        aclient.get_user_task_list_gid(0, True)
-    assert caplog.record_tuples == [
-            ('asana_extensions.asana.client', logging.ERROR,
-                "Could not find requested data in get_user_task_list_gid():"
-                + " Not Found"),
-    ]
+            raise_asana_error)
+    subtest_asana_error_handler_func(caplog, asana.error.ServerError, 0,
+            aclient.get_user_task_list_gid, 0, True)
 
 
 
+@pytest.mark.asana_error_data.with_args(asana.error.InvalidRequestError)
 def test_get_section_gids_in_project_or_utl(monkeypatch, caplog, project_test,
-        section_in_project_test, raise_no_authorization_error):
+        sections_in_project_test, raise_asana_error):
     """
     Tests the `get_section_gids_in_project_or_utl()` method.
 
@@ -511,6 +555,9 @@ def test_get_section_gids_in_project_or_utl(monkeypatch, caplog, project_test,
     # pylint: disable=no-member     # asana.Client dynamically adds attrs
     caplog.set_level(logging.ERROR)
 
+    # Only need 1 section
+    section_in_project_test = sections_in_project_test[0]
+
     try:
         sect_gids = aclient.get_section_gids_in_project_or_utl(
                 project_test['gid'])
@@ -522,16 +569,30 @@ def test_get_section_gids_in_project_or_utl(monkeypatch, caplog, project_test,
 
     assert int(section_in_project_test['gid']) in sect_gids
 
-    # Need to monkeypatch cached client since class dynamically creates attrs
+    # Function-specific practical test of @asana_error_handler
     client = aclient._get_client()
+    # Need to monkeypatch cached client since class dynamically creates attrs
     monkeypatch.setattr(client.sections, 'get_sections_for_project',
-            raise_no_authorization_error)
+            raise_asana_error)
+    subtest_asana_error_handler_func(caplog, asana.error.InvalidRequestError, 0,
+            aclient.get_section_gids_in_project_or_utl, project_test['gid'])
 
-    caplog.clear()
-    with pytest.raises(asana.error.NoAuthorizationError):
-        aclient.get_section_gids_in_project_or_utl(project_test['gid'])
-    assert caplog.record_tuples == [
-            ('asana_extensions.asana.client', logging.ERROR,
-                "Failed to access API in get_section_gids_in_project_or_utl() -"
-                + " Not Authorized: No Authorization"),
-    ]
+
+
+def test_pagination(project_test, sections_in_project_test):
+    """
+    Tests compatibility with `asana` package to ensure that any pagination is
+    handled in a way that is compatible with how this project expects it.
+
+    ** Consumes at least 2 API calls. **
+    (varies depending on data size, but only 2 calls intended)
+    """
+    client = aclient._get_client()
+    client.options['page_size'] = 1
+
+    sect_gids = aclient.get_section_gids_in_project_or_utl(project_test['gid'])
+
+    # Should match exactly, but other tests may have added more sects to server
+    assert len(sect_gids) >= len(sections_in_project_test)
+    for sect in sections_in_project_test:
+        assert int(sect['gid']) in sect_gids